Select video profile by ID only
Do not apply the resolution constraints when searching for a video
profile by unique ID, as the video profile ID is unique anyway so this
is unnecessary work.

Instead, apply constraints while listing supported resolutions, such
that only the resolutions matching the current constraints are returned
and can be selected by the internal WebRTC implementation.

TODO: Note that currently this returns resolutions not in the video profile.