Ensure consistency between WebRTC and MediaCapture
Ensure that the media type selected by the driver from the MediaCapture
settings is always the same as the video format selected by WebRTC.

The potential discrepancy arises from the fact that setting the
RecordMediaDescription field on the MediaCapture init settins tells the
driver to use that media type, but WebRTC doesn't know anything about
thsi choice and would use a different heuristic to select a video
format.

To prevent that, if a video profile is selected, only return a single
supported video format corresponding to the one selected in the
MediaCapture object.