Wait for StopCapture() to finish before returning (#28)
Ensure `StopCapture()` finished stopping and cleaning up the capture
device before it returns. Otherwise subsequent calls to the media
capture device might proceed to access it in parallel of that work,
creating a race condition.

Because `VideoCapturer` uses a single lock to protect access to the
`CaptureDevice` member as well as all its other members, this lock
probably needs to be held during the entire time of the per-frame
callback. And because it is the source of the deadlock, it needs to be
acquired only when it is safe to do so, that is not when the capture
is stopping. So this change adds a new lock which is only held by the
per-frame callback during its work, and briefly at the beginning and
end of capture shutdown. This way, with the `is_shutting_down_` flag,
the per-frame callback can avoid taking the API lock if another thread
informed it that capture is being shut down, and not block the media
capture pipeline from shutting down by returning immediately.

This fixes #189.