Ignore exceptions when closing video capture
When `CaptureDevice::Cleanup()` is called, it attempts to stop the local
video recording calling `IMediaCapture::StopRecordAsync()`. Sometimes
this fails with `MF_E_SAMPLEALLOCATOR_EMPTY`, often when the system load
is high and the recording does not stop fast enough while frames are
still produced, but not consumed.

Because the code uses `IAsyncAction::get()` to do a blocking call and
wait for `StopRecordAsync()` to return, and because `get()` has will
throw an exception on error, the result is that some exception may be
thrown although there is fundamentally no error and the recording is
stopped as expected.

This change ignores all exceptions thrown by `get()` during the call to
`StopRecordAsync()` to prevent the application from crashing with an
uncaught exception.